Deprecation of Microsoft Teams Classic
Due to the recent deprecation of Microsoft Teams Classic and the renaming of the work or school version back to "Microsoft Teams," the script encountered an issue where both packages were named "MicrosoftTeams.pkg." This caused each definition to appear corrupted, leading to the removal of either the older or newer version and triggering a constant redownload of the Microsoft Teams packages.

Since the Classic version is no longer in use, I decided to remove all references to Microsoft Teams Classic and reorganized the array numbering accordingly.

Version updated to 3.8.